指派订单+bug

+++ ant-api/src/main/java/com/bx/ant/pageModel/DeliverOrderQuery.java
@@ -19,7 +19,7 @@ public class DeliverOrderQuery extends DeliverOrder {
     private double amountElement;
     private String createDate;
     private String requiredDate;
-    private double time;
+    private Integer time;
 
     public String getStatusName() {
         return ConvertNameUtil.getString(this.statusName);
@@ -109,11 +109,11 @@ public class DeliverOrderQuery extends DeliverOrder {
         this.requiredDate = requiredDate;
     }
 
-    public Double getTime() {
+    public Integer getTime() {
         return time;
     }
 
-    public void setTime(double time) {
+    public void setTime(Integer time) {
         this.time = time;
     }
 }
